// TryHackSwigger-main/src/lib/types.ts
export enum BlueLabCategory {
  LOG_ANALYSIS = 'Log Analysis / SIEM',
  INCIDENT_RESPONSE = 'Incident Response',
  DIGITAL_FORENSICS = 'Digital Forensics',
  THREAT_HUNTING = 'Threat Hunting',
  NETWORK_DEFENSE = 'Network Defense / SOC',
  MALWARE_ANALYSIS = 'Malware Analysis'
}

export enum LabType {
  GUIDED = 'Guided',
  SCENARIO = 'Scenario'
}

export enum TeamType {
  RED = 'RED',
  BLUE = 'BLUE',
  PURPLE = 'PURPLE'
}

export enum Difficulty {
  EASY = 'Easy',
  MEDIUM = 'Medium',
  HARD = 'Hard',
  INSANE = 'Insane'
}